### 2021-03-14 #3111 @whedon
[PRE REVIEW]: statsExpressions: R Package for Tidy Dataframes and Expressions with Statistical Details
**Submitting author:** @IndrajeetPatil (<a href="http://orcid.org/0000-0003-1995-6531">Indrajeet Patil</a>)
**Repository:** <a href="https://github.com/IndrajeetPatil/statsExpressions" target ="_blank">https://github.com/IndrajeetPatil/statsExpressions</a>
**Version:** 1.0.0
