### 2021-03-14 #3111 @arfon
@IndrajeetPatil - this PR should fix your paper compilation https://github.com/IndrajeetPatil/statsExpressions/pull/81
