+++ tests/testthat/test-oneway_anova_nonparametric.R
@@ -136,12 +136,12 @@ test_that(
           ", ",
           widehat(italic("W"))["Kendall"],
           " = ",
-          "0.6021",
+          "0.1750",
           ", CI"["99%"],
           " [",
-          "0.6021",
+          "0.1146",
           ", ",
-          "0.9748",
+          "0.2793",
           "]",
           ", ",
           italic("n")["pairs"],
@@ -183,12 +183,12 @@ test_that(
           ", ",
           widehat(italic("W"))["Kendall"],
           " = ",
-          "0.484",
+          "0.911",
           ", CI"["90%"],
           " [",
-          "0.343",
+          "0.905",
           ", ",
-          "0.969",
+          "0.918",
           "]",
           ", ",
           italic("n")["pairs"],
