+++ DESCRIPTION
@@ -28,7 +28,7 @@ Imports:
     BayesFactor (>= 0.9.12-4.2),
     correlation (>= 0.6.0),
     dplyr,
-    effectsize (>= 0.4.3),
+    effectsize (>= 0.4.4),
     insight (>= 0.13.0),
     ipmisc (>= 6.0.0),
     parameters (>= 0.12.0),
@@ -50,7 +50,8 @@ Suggests:
     testthat,
     utils
 Remotes:
-    richarddmorey/BayesFactor/pkg/BayesFactor
+    richarddmorey/BayesFactor/pkg/BayesFactor,
+    easystats/effectsize
 VignetteBuilder: 
     knitr
 Config/testthat/edition: 3
