+++ NAMESPACE
@@ -12,12 +12,6 @@ export(bf_extractor)
 export(contingency_table)
 export(corr_test)
 export(correlation)
-export(expr_contingency_tab)
-export(expr_corr_test)
-export(expr_meta_random)
-export(expr_oneway_anova)
-export(expr_t_onesample)
-export(expr_t_twosample)
 export(expr_template)
 export(meta_analysis)
 export(one_sample_test)
