+++ NAMESPACE
@@ -56,6 +56,7 @@ importFrom(effectsize,kendalls_w)
 importFrom(effectsize,omega_squared)
 importFrom(effectsize,rank_biserial)
 importFrom(effectsize,rank_epsilon_squared)
+importFrom(insight,format_value)
 importFrom(insight,standardize_names)
 importFrom(ipmisc,"%$%")
 importFrom(ipmisc,"%<-%")
