+++ tests/testthat/test-oneway_anova_bayes.R
@@ -1,242 +1,242 @@
-if (packageVersion("parameters") > "0.11.0") {
+# bayes factor (between-subjects - anova) ---------------------------------
 
-  # bayes factor (between-subjects - anova) ---------------------------------
+test_that(
+  desc = "bayes factor (between-subjects - anova)",
+  code = {
+    skip_if(getRversion() < "4.0")
+    skip_on_cran()
 
-  test_that(
-    desc = "bayes factor (between-subjects - anova)",
-    code = {
-      skip_if(getRversion() < "4.0")
+    # extracting results from where this function is implemented
+    set.seed(123)
+    df_results <-
+      oneway_anova(
+        type = "bayes",
+        data = ggplot2::msleep,
+        x = "vore",
+        y = brainwt,
+        bf.prior = 0.99
+      )
+
+    # extracting expr
+    set.seed(123)
+    results <-
+      oneway_anova(
+        type = "bayes",
+        data = ggplot2::msleep,
+        x = vore,
+        y = "brainwt",
+        bf.prior = 0.88,
+        k = 2 # don't change; tests fail on Ubuntu otherwise
+      )
+
+    # check bayes factor values
+    expect_equal(df_results$bf10[[1]], 0.1177186, tolerance = 0.001)
+
+    # call
+    expect_identical(
+      results$expression[[1]],
+      ggplot2::expr(
+        paste(
+          "log"["e"] * "(BF"["01"] * ") = " * "1.92" * ", ",
+          widehat(italic(R^"2"))["Bayesian"]^"posterior" * " = " * "0.00" * ", ",
+          "CI"["95%"]^"HDI" * " [" * "0.00" * ", " * "0.08" * "], ",
+          italic("r")["Cauchy"]^"JZS" * " = " * "0.88"
+        )
+      )
+    )
+
+    # data where it works
+    set.seed(123)
+    results2 <-
+      oneway_anova(
+        type = "bayes",
+        data = iris,
+        x = Species,
+        y = Sepal.Length,
+        conf.level = 0.99,
+        conf.method = "eti",
+        k = 2 # don't change; tests fail on Ubuntu otherwise
+      )
+
+    expect_identical(
+      results2$expression[[1]],
+      ggplot2::expr(
+        paste(
+          "log"["e"] * "(BF"["01"] * ") = " * "-65.10" * ", ",
+          widehat(italic(R^"2"))["Bayesian"]^"posterior" * " = " * "0.61" * ", ",
+          "CI"["99%"]^"HDI" * " [" * "0.51" * ", " * "0.68" * "], ",
+          italic("r")["Cauchy"]^"JZS" * " = " * "0.71"
+        )
+      )
+    )
+  }
+)
+
+# bayes factor (within-subjects - anova) ---------------------------------
+
+test_that(
+  desc = "bayes factor (within-subjects - anova)",
+  code = {
+    skip_on_cran()
+
+    if (utils::packageVersion("BayesFactor") >= package_version("0.9.12-4.3")) {
+
+      # dataframe
+      dat <-
+        structure(list(Taste = c(
+          5.4, 5.5, 5.55, 5.85, 5.7, 5.75, 5.2,
+          5.6, 5.5, 5.55, 5.5, 5.4, 5.9, 5.85, 5.7, 5.45, 5.55, 5.6, 5.4,
+          5.4, 5.35, 5.45, 5.5, 5.35, 5.25, 5.15, 5, 5.85, 5.8, 5.7, 5.25,
[…]
