+++ tests/testthat/test-corr_test.R
@@ -214,93 +214,90 @@ test_that(
 )
 
 
-if (packageVersion("parameters") > "0.11.0") {
+# bayes factor (correlation test) --------------------------------------
 
-  # bayes factor (correlation test) --------------------------------------
+test_that(
+  desc = "bayes factor (correlation test) - without NAs",
+  code = {
+    skip_if(getRversion() < "4.0")
 
-  test_that(
-    desc = "bayes factor (correlation test) - without NAs",
-    code = {
-      skip_if(getRversion() < "4.0")
+    # extracting results from where this function is implemented
+    set.seed(123)
+    df <-
+      corr_test(
+        type = "bayes",
+        data = iris,
+        y = Sepal.Length,
+        x = "Sepal.Width"
+      )
 
-      # extracting results from where this function is implemented
-      set.seed(123)
-      df <-
-        corr_test(
-          type = "bayes",
-          data = iris,
-          y = Sepal.Length,
-          x = "Sepal.Width"
-        )
+    # check bayes factor values
+    expect_equal(df$bf10, 0.3445379, tolerance = 0.001)
 
-      # check bayes factor values
-      expect_equal(df$bf10, 0.3445379, tolerance = 0.001)
+    set.seed(123)
+    subtitle1 <-
+      corr_test(
+        type = "bayes",
+        data = iris,
+        y = Sepal.Length,
+        x = Sepal.Width,
+        top.text = "huh"
+      )
 
-      set.seed(123)
-      subtitle1 <-
-        corr_test(
-          type = "bayes",
-          data = iris,
-          y = Sepal.Length,
-          x = Sepal.Width,
-          top.text = "huh"
-        )
+    expect_identical(
+      subtitle1$expression[[1]],
+      ggplot2::expr(
+        atop(displaystyle("huh"), expr = paste(
+          "log"["e"] * "(BF"["01"] * ") = " * "1.07" * ", ",
+          widehat(rho)["Pearson"]^"posterior" * " = " * "-0.12" * ", ",
+          "CI"["95%"]^"HDI" * " [" * "-0.24" * ", " * "0.02" * "], ",
+          italic("r")["beta"]^"JZS" * " = " * "1.41"
+        ))
+      )
+    )
+  }
+)
 
-      expect_identical(
-        subtitle1$expression[[1]],
-        ggplot2::expr(
-          atop(displaystyle("huh"), expr = paste(
-            "log"["e"] * "(BF"["01"] * ") = " * "1.07" * ", ",
-            widehat(rho)["Pearson"]^"posterior" * " = " * "-0.12" * ", ",
-            "CI"["95%"]^"HDI" * " [" * "-0.24" * ", " * "0.02" * "], ",
-            italic("r")["beta"]^"JZS" * " = " * "1.41"
-          ))
+test_that(
+  desc = "bayes factor (correlation test) - with NAs",
+  code = {
+    skip_if(getRversion() < "4.0")
+
+    # extracting results from where this function is implemented
+    set.seed(123)
+    df <-
+      corr_test(
+        type = "bayes",
+        data = ggplot2::msleep,
+        y = names(ggplot2::msleep)[10],
+        x = "sleep_rem",
+      )
+
+    # check bayes factor values
+    expect_equal(df$bf10, 0.6539296, tolerance = 0.001)
+
+    set.seed(123)
+    subtitle1 <-
+      corr_test(
+        type = "bayes",
+        data = ggplot2::msleep,
+        y = brainwt,
+        x = sleep_rem,
+        bf.prior = 0.8,
+        conf.level = 0.99,
+        k = 3
+      )
+
+    expect_identical(
[…]
