+++ DESCRIPTION
@@ -30,13 +30,13 @@ Imports:
     dplyr,
     effectsize (>= 0.4.4),
     insight (>= 0.13.0),
-    ipmisc (>= 6.0.0),
+    ipmisc,
     parameters (>= 0.12.0),
     performance,
     rlang,
     stats,
     tidyr,
-    WRS2 (>= 1.1-1)
+    WRS2
 Suggests:
     afex,
     ggplot2,

[…]
