+++ cran-comments.md
@@ -10,7 +10,8 @@
 
 0 errors | 0 warnings | 0 note
 
-  - Fixes failing tests due to updates to `insight` and `parameters` packages.
+  - Fixes failing tests due to updates to `effectsize` packages.
+  - Fixes failing test in additional checks.
 
 ## revdepcheck results
 
@@ -20,4 +21,3 @@ across CRAN and dev versions of this package.
  * We saw 0 new problems
 
  * We failed to check 0 packages
-
