+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: statsExpressions
 Title: Tidy Dataframes and Expressions with Statistical Details
-Version: 1.0.0.9000
+Version: 1.0.1
 Authors@R: 
     person(given = "Indrajeet",
            family = "Patil",
@@ -26,13 +26,13 @@ Depends:
     R (>= 3.6.0)
 Imports:
     BayesFactor (>= 0.9.12-4.2),
-    correlation (>= 0.6.0),
+    correlation (>= 0.6.1),
     dplyr,
-    effectsize (>= 0.4.4),
+    effectsize (>= 0.4.4-1),
     insight (>= 0.13.0),
     ipmisc,
-    parameters (>= 0.12.0),
-    performance,
+    parameters (>= 0.13.0),
+    performance (>= 0.7.1),
     rlang,
     stats,
     tidyr,
@@ -49,9 +49,6 @@ Suggests:
     spelling,
     testthat,
     utils
-Remotes:
-    richarddmorey/BayesFactor/pkg/BayesFactor,
-    easystats/effectsize
 VignetteBuilder: 
     knitr
 Config/testthat/edition: 3
