+++ R/meta_analysis.R
@@ -15,7 +15,7 @@
 #'
 #' @description
 #'
-#' \Sexpr[results=rd, stage=render]{rlang:::lifecycle("stable")}
+#' 
 #'
 #' A dataframe containing results from random-effects meta-analysis.
 #'
