+++ R/helpers_expr_templates.R
@@ -3,8 +3,6 @@
 #'
 #' @description
 #'
-#' \Sexpr[results=rd, stage=render]{rlang:::lifecycle("questioning")}
-#'
 #' Creates an expression from a dataframe containing statistical details.
 #' Ideally, this dataframe would come from having run `tidy_model_parameters`
 #' function on your model object.
