+++ man/contingency_table.Rd
@@ -85,8 +85,6 @@ mostly relevant in the context of \code{ggstatsplot} functions.}
 \item{...}{Additional arguments (currently ignored).}
 }
 \description{
-\Sexpr[results=rd, stage=render]{rlang:::lifecycle("stable")}
-
 A dataframe containing results from for contingency table analysis or
 goodness of fit test.
 
