+++ man/corr_test.Rd
@@ -62,8 +62,6 @@ mostly relevant in the context of \code{ggstatsplot} functions.}
 \item{...}{Additional arguments (currently ignored).}
 }
 \description{
-\Sexpr[results=rd, stage=render]{rlang:::lifecycle("stable")}
-
 A dataframe containing results from correlation test with confidence
 intervals for the correlation coefficient estimate. Results are extracted
 via \code{correlation::correlation}.
