### 2021-04-12 #3111 @IndrajeetPatil
Can this submission be reconsidered for review?

Two of my other submissions (both `R` packages) at _JOSS_ were submitted after this submission and are already under review. So I am guessing someone might be available now to edit this submission?
