### 2021-04-12 #3111 @arfon
> Two of my other submissions (both `R` packages) at JOSS were submitted after this submission and are already under review. So I am guessing someone might be available now to edit this submission?

Not necessarily so. Please be patient.
