+++ R/corr_test.R
@@ -3,7 +3,7 @@
 #'
 #' @description
 #'
-#' 
+#'
 #'
 #'  A dataframe containing results from correlation test with confidence
 #'  intervals for the correlation coefficient estimate. Results are extracted
