+++ R/meta_analysis.R
@@ -15,7 +15,7 @@
 #'
 #' @description
 #'
-#' 
+#'
 #'
 #' A dataframe containing results from random-effects meta-analysis.
 #'
