### 2021-04-15 a7378ee
badges in README; bump to devel
