+++ NEWS.md
@@ -1,3 +1,5 @@
+# statsExpressions 1.0.1.9000
+
 # statsExpressions 1.0.1
 
 BREAKING CHANGES
