+++ R/two_sample_test.R
@@ -19,8 +19,6 @@
 #'
 #' @description
 #'
-#'
-#'
 #'  A dataframe containing details from results of a two-sample test and effect
 #'  size plus confidence intervals.
 #'
