+++ tests/testthat/test-loop_execution_bayes.R
@@ -4,7 +4,7 @@
 test_that(
   desc = "corr_test works in loop",
   code = {
-    skip_on_cran()
+    # skip_on_cran()
     col.name <- colnames(mtcars)
 
     set.seed(123)
@@ -27,7 +27,7 @@ test_that(
 test_that(
   desc = "contingency_table works in loop",
   code = {
-    skip_on_cran()
+    # skip_on_cran()
     df <- dplyr::select(mtcars, am, cyl, vs)
     col.name <- colnames(df)
 
@@ -50,7 +50,7 @@ test_that(
 test_that(
   desc = "expr_ttest works in loop",
   code = {
-    skip_on_cran()
+    # skip_on_cran()
 
     # working with loops
     df <- dplyr::select(mtcars, am, wt, mpg)
@@ -89,7 +89,7 @@ test_that(
 test_that(
   desc = "expr_anova works in loop",
   code = {
-    skip_on_cran()
+    # skip_on_cran()
 
     # working with loops
     df <- dplyr::select(mtcars, cyl, wt, mpg)
