+++ R/corr_test.R
@@ -3,8 +3,6 @@
 #'
 #' @description
 #'
-#'
-#'
 #'  A dataframe containing results from correlation test with confidence
 #'  intervals for the correlation coefficient estimate. Results are extracted
 #'  via `correlation::correlation`.
