### 2021-04-23 #3111 @IndrajeetPatil
`statsExpressions` has wildly different goals than `broom`.

- `broom`: convert regression and other modeling **objects** to dataframes
- `statsExpressions`: provide consistent syntax to run different tests and extract dataframes and expressions

So `statsExpressions` doesn't work with objects, but rather provides functions to carry out different types of analyses.

Additionally, it is more opinionated than `broom`. For example, `broom` won't add any other information than present in underlying objects (e.g., it doesn't add effect sizes and their confidence intervals), while `statsExpressions` does.

``` r
broom::tidy(t.test(mtcars$wt, mu = 3))
#> # A tibble: 1 x 8
#>   estimate statistic p.value parameter conf.low conf.high method     alternative
#>      <dbl>     <dbl>   <dbl>     <dbl>    <dbl>     <dbl> <chr>      <chr>
#> 1     3.22      1.26   0.218        31     2.86      3.57 One Sampl… two.sided

(df <- statsExpressions::one_sample_test(mtcars, wt, test.value = 3))
#> # A tibble: 1 x 11
#>      mu statistic df.error p.value method estimate conf.level conf.low conf.high
#>   <dbl>     <dbl>    <dbl>   <dbl> <chr>     <dbl>      <dbl>    <dbl>     <dbl>
#> 1     3      1.26       31   0.218 One S…    0.217       0.95   -0.129     0.566
#> # … with 2 more variables: effectsize <chr>, expression <list>
```

Additionally, `statsExpressions` provides dataframes containing expressions with statistical details. `broom` does no such thing.

```r
library(ggplot2)

ggplot(mtcars, aes(wt)) + geom_histogram()  +
  labs(subtitle = df$expression[[1]])
#> `stat_bin()` using `bins = 30`. Pick better value with `binwidth`.
```

![](https://i.imgur.com/ahSKR48.png)
