### 2021-04-23 #3111 @mikldk
@IndrajeetPatil Thanks. I think it would be appropriate if `broom` was briefly mentioned in the paper and how this package differs from it.

A minor question from your example just given:

``` r
broom::tidy(t.test(mtcars$wt, mu = 3))
#> # A tibble: 1 x 8
#>   estimate statistic p.value parameter conf.low conf.high method     alternative
#>      <dbl>     <dbl>   <dbl>     <dbl>    <dbl>     <dbl> <chr>      <chr>
#> 1     3.22      1.26   0.218        31     2.86      3.57 One Sampl… two.sided

(df <- statsExpressions::one_sample_test(mtcars, wt, test.value = 3))
#> # A tibble: 1 x 11
#>      mu statistic df.error p.value method estimate conf.level conf.low conf.high
#>   <dbl>     <dbl>    <dbl>   <dbl> <chr>     <dbl>      <dbl>    <dbl>     <dbl>
#> 1     3      1.26       31   0.218 One S…    0.217       0.95   -0.129     0.566
#> # … with 2 more variables: effectsize <chr>, expression <list>
```

Wouldn't you expect that `one_sample_test()`'s `conf.low` / `conf.high` were centered around `mu` and not 0 (like `t.test()`)?
