### 2021-04-23 #3111 @IndrajeetPatil
> I think it would be appropriate if broom was briefly mentioned in the paper and how this package differs from it.

Sure, I can do that.

> Wouldn't you expect that one_sample_test()'s conf.low / conf.high were centered around mu and not 0 (like t.test())?

The estimate is actually for Hedge's _g_, which you couldn't see since not all columns of the dataframe were printed.

``` r
options(tibble.width = Inf)
(df <- statsExpressions::one_sample_test(mtcars, wt, test.value = 3))
#> # A tibble: 1 x 11
#>      mu statistic df.error p.value method            estimate conf.level
#>   <dbl>     <dbl>    <dbl>   <dbl> <chr>                <dbl>      <dbl>
#> 1     3      1.26       31   0.218 One Sample t-test    0.217       0.95
#>   conf.low conf.high effectsize expression
#>      <dbl>     <dbl> <chr>      <list>
#> 1   -0.129     0.566 Hedges' g  <language>
```
