### 2021-04-23 IndrajeetPatil/statsExpressions#89 @IndrajeetPatil
add `broom` citation to paper
add `broom` citation to paper (cf. https://github.com/openjournals/joss-reviews/issues/3111#issuecomment-825586441)
