### 2021-04-23 #3111 @mikldk
> The estimate is actually for Hedge's _g_, which you couldn't see since not all columns of the dataframe were printed.

Okay. Is it possible to also get `\hat{mu}` and confidence interval for `mu` similar to what `t.test()` outputs?
