### 2021-04-23 #3111 @IndrajeetPatil
>  Is it possible to also get \hat{mu} and confidence interval for mu similar to what t.test() outputs?

Currently, no. The user can either get Hedge's _g_ or Cohen's _d_ for this test. If you want, I can add to the feature request list.

Usually, across different `type` arguments, there will just be one type of estimate available.

``` r
library(statsExpressions)
options(tibble.width = Inf)

one_sample_test(mtcars, wt, test.value = 3, type = "parametric")
#> # A tibble: 1 x 11
#>      mu statistic df.error p.value method            estimate conf.level
#>   <dbl>     <dbl>    <dbl>   <dbl> <chr>                <dbl>      <dbl>
#> 1     3      1.26       31   0.218 One Sample t-test    0.217       0.95
#>   conf.low conf.high effectsize expression
#>      <dbl>     <dbl> <chr>      <list>
#> 1   -0.129     0.566 Hedges' g  <language>

one_sample_test(mtcars, wt, test.value = 3, type = "nonparametric")
#> # A tibble: 1 x 9
#>   statistic p.value method                    estimate conf.level conf.low
#>       <dbl>   <dbl> <chr>                        <dbl>      <dbl>    <dbl>
#> 1      5.77   0.308 Wilcoxon signed rank test    0.208       0.95   -0.249
#>   conf.high effectsize        expression
#>       <dbl> <chr>             <list>
#> 1     0.578 r (rank biserial) <language>

one_sample_test(mtcars, wt, test.value = 3, type = "robust")
#> # A tibble: 1 x 9
#>   statistic p.value method                                 estimate conf.low
#>       <dbl>   <dbl> <chr>                                     <dbl>    <dbl>
#> 1      1.18    0.24 Bootstrap-t method for one-sample test     3.20     2.81
#>   conf.high conf.level effectsize   expression
#>       <dbl>      <dbl> <chr>        <list>
#> 1      3.58       0.95 Trimmed mean <language>

one_sample_test(mtcars, wt, test.value = 3, type = "bayesian")
#> # A tibble: 2 x 13
#>   term       estimate conf.level conf.low conf.high    pd rope.percentage
#>   <chr>         <dbl>      <dbl>    <dbl>     <dbl> <dbl>           <dbl>
#> 1 Difference   -0.192       0.95   -0.534     0.173 0.859           0.274
#> 2 Cohens_d      3.19        0.95    2.29      4.06  1               0
#>   prior.distribution prior.location prior.scale  bf10 method          expression
#>   <chr>                       <dbl>       <dbl> <dbl> <chr>           <list>
#> 1 cauchy                          0       0.707 0.387 Bayesian t-test <language>
#> 2 cauchy                          0       0.707 0.387 Bayesian t-test <language>
```
