### 2021-05-03 #3111 @humanfactors
@mikldk I would like to flag a potential perceived conflict of interest (i.e., to "avoid the appearance of conflicts of interest" as per JOSS policy). Both @IndrajeetPatil and I are members of the [easystats](https://easystats.github.io/blog/portfolio/) contributor list and we both were CRAN co-authors on the `bayestestR` package (https://cran.r-project.org/web/packages/bayestestR/index.html).

If this is perceived as a possible COI, I would request this COI is waived on the basis that:
- Indrajeet Patil and I have never co-authored a published manuscript, and do not have any prior or current collaborations related to publications. Nor do I see there to be any financial or academic benefits to me from this work that would prevent me from giving an impartial assessment of the work.
- We are not personally or professionally antiquated
- This project does not directly add to the easystats ecosystem

I place this here to be as thorough as possible and ensure I conform to policies.
