### 2021-05-03 #3236 @whedon
[REVIEW]: statsExpressions: R Package for Tidy Dataframes and Expressions with Statistical Details
**Submitting author:** @IndrajeetPatil (<a href="http://orcid.org/0000-0003-1995-6531">Indrajeet Patil</a>)
**Repository:** <a href="https://github.com/IndrajeetPatil/statsExpressions" target ="_blank">https://github.com/IndrajeetPatil/statsExpressions</a>
**Version:** v1.0.1
