### 2021-05-03 #3236 @mikldk
@humanfactors, @tomfaulkenberry: Thanks for agreeing to review. Please carry out your review in this issue by updating the checklist above and giving feedback in this issue. The reviewer guidelines are available here: <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html>. If possible create issues (and cross-reference) in the submission's repository to avoid too specific discussions in this review thread.

@IndrajeetPatil Please note the error above "Failed to discover a `Statement of need` section in paper".
