### 2021-05-03 7690a25
closes #89; adds Statement of Need; add missing doi
