+++ paper/paper.bib
@@ -13,7 +13,7 @@
 
 
 @Article{Lüdecke2020parameters,
-    title = {parameters: Extracting, Computing and Exploring the Parameters of Statistical Models using {R}.},
+    title = {{parameters}: Extracting, Computing and Exploring the Parameters of Statistical Models using {R}.},
     volume = {5},
     doi = {10.21105/joss.02445},
     number = {53},
@@ -23,20 +23,20 @@
     pages = {2445},
   }
 
-
 @Article{Lüdecke2020performance,
-    title = {performance: Assessment of Regression Models Performance},
-    author = {Daniel Lüdecke and Dominique Makowski and Philip Waggoner and Indrajeet Patil},
-    journal = {CRAN},
-    year = {2020},
-    note = {R package},
-    doi = {10.5281/zenodo.3952174},
-    url = {https://easystats.github.io/performance/},
+    title = {{performance}: An {R} Package for Assessment, Comparison and Testing of Statistical Models},
+    author = {Daniel Lüdecke and Mattan S. Ben-Shachar and Indrajeet Patil and Philip Waggoner and Dominique Makowski},
+    year = {2021},
+    journal = {Journal of Open Source Software},
+    volume = {6},
+    number = {60},
+    pages = {3139},
+    doi = {10.21105/joss.03139},
   }
 
 
 @Article{Lüdecke2019,
-    title = {insight: A Unified Interface to Access Information from Model Objects in R.},
+    title = {{insight}: A Unified Interface to Access Information from Model Objects in {R}.},
     volume = {4},
     doi = {10.21105/joss.01412},
     number = {38},
@@ -59,16 +59,14 @@
     url = {https://joss.theoj.org/papers/10.21105/joss.02306},
   }
 
-
-@Article{Patil2018,
-    title = {{ggstatsplot}: 'ggplot2' Based Plots with Statistical Details},
+@Article{Patil2021,
+    title = {Visualizations with statistical details: The 'ggstatsplot' approach},
     author = {Indrajeet Patil},
-    year = {2018},
-    journal = {CRAN},
-    url = {https://CRAN.R-project.org/package=ggstatsplot},
-    doi = {10.5281/zenodo.2074621},
+    year = {2021},
+    journal = {PsyArxiv},
+    url = {https://psyarxiv.com/p7mku/},
+    doi = {10.31234/osf.io/p7mku},
   }
-
 
 @Manual{Morey2020,
     title = {BayesFactor: Computation of Bayes Factors for Common Designs},
@@ -85,6 +83,7 @@
     journal = {Behavior Research Methods},
     volume = {52},
     pages = {464--488},
+    doi = {10.3758/s13428-019-01246-w}
   }
 
 @Article{Makowski2019,
@@ -121,12 +120,21 @@
   }
 
 
+@Manual{Robinson2021,
+    title = {broom: Convert Statistical Objects into Tidy Tibbles},
+    author = {David Robinson and Alex Hayes and Simon Couch},
+    year = {2021},
+    note = {R package version 0.7.6},
+    url = {https://CRAN.R-project.org/package=broom},
+  }
+
 @article{van2020jasp,
   title={The JASP guidelines for conducting and reporting a Bayesian analysis},
   author={van Doorn, Johnny and van den Bergh, Don and B{\"o}hm, Udo and Dablander, Fabian and Derks, Koen and Draws, Tim and Etz, Alexander and Evans, Nathan J and Gronau, Quentin F and Haaf, Julia M and others},
   journal={Psychonomic Bulletin \& Review},
   pages={1--14},
   year={2020},
+  doi={10.3758/s13423-020-01798-5},
   publisher={Springer}
 }
 
[…]
