### 2021-05-03 #3236 @IndrajeetPatil
> @IndrajeetPatil Please note the error above "Failed to discover a Statement of need section in paper".

@mikldk Thanks, I have renamed the relevant section.

I have also added a sentence about how the software differs from `broom`, like you had asked:

https://github.com/IndrajeetPatil/statsExpressions/blob/7690a2587a2d856cf970002808da2a0bf13d8982/paper/paper.Rmd#L117-L120
