### 2021-05-05 IndrajeetPatil/statsExpressions#93 @humanfactors
[JOSS REVIEW] Main feedback
**Submission:** https://github.com/openjournals/joss-reviews/issues/3236
**Title:** statsExpressions: R Package for Tidy Dataframes and Expressions with Statistical Details
**Review Outcome**: Minor Revisions

## Overview

`statsExpressions` is an R package for quickly running various forms of statistical tests using a single framework (e.g.,
consistent syntax) and offers users "pre-formatted statistical expressions" for annotating plots, as well as consistent
dataframe outputs. The package covers a comprehensive range of commonly used statistical tests (at least in the
behavioral sciences).

I commend the author on their effort, the package appears to be well written and documented, with comprehensive tests and exhaustive use of continuous integration. I installed the software successfully, and ran a few statistical tests myself. I could see myself using this software in the future.

At this stage, there are a number of relatively minor revisions that should be made to the package and manuscript. Below, I list the changes which would allow me to recommend acceptance.

## Manuscript and Documentation

I've made some minor typographical corrections and commentary to the manuscript in [this attached PDF (10.21105.joss.03236_WithComments.pdf)](https://github.com/IndrajeetPatil/statsExpressions/files/6425382/10.21105.joss.03236_WithComments.pdf). Below are more general or substantial issues to be addressed.

### Package summary and need statement

There is currently inconsistency across sources regarding the exact purpose and scope of the package which needs to be resolved.

> \[Statement of Need\] The aim of this package is to provide an approachable and intuitive syntax to carry out common statistical tests across diverse statistical approaches.

> \[Summary\] The statsExpressions package is designed to facilitate producing dataframes with rich statistical details for the most common types of statistical approaches and tests

> Although the primary focus of this package is to get expressions containing statistical results, one can also use it to extract dataframes containing these details. <https://indrajeetpatil.github.io/statsExpressions/articles/stats_details.html>

These various sources emphasize somewhat different package features. For instance, the statement of need does not
mention expressions or consistent dataframe outputs, while the vignette description downplays what appears to be a key selling point of the package: consistent syntax and outputs.

There is a need for these various descriptions to become more homogeneous (consistent) with each other. This will
require updating in documentation also (README and Vignettes). I foresee that a hybrid of all three descriptions that clearly explains the purpose of the package is warranted.

Note that I have seen mixed answers regarding whether an explicit statement of need is required if it is already
covered in the summary (in the case you experience a lot of overlap). @mikldk

You should mention the specific user-base (target audience) this package would appeal to in the statement of need (if applicable).

I also want to flag that it was not immediately clear to me what an "expression" was, and I initially had to read
further into the README to know what the term "expression" was referring to. I recommend placing some added description in parentheses when first introducing the term expression (in both manuscript and README). For instance, "_expression (i.e., a pre-formatted in-text statistical result)_".

### Comparison to Other Packages

I had a few issues with the contents of this section which require revision.

1.  Only the last paragraph (L43-46) is directly discussing comparisons to other packages. The remaining content could be moved into a separate subsection, or alternatively, have a new header (e.g., "Package overview").

2.  The language in L23-36 is somewhat speculative and encroaching informality in tone, particularly from L32. I suggest revising for objective tone, noting the following:
      - L23-31: I believe providing a single concrete example would be powerful. For instance, of the packages/functions `statsExpressions` covers, which of these have an object that return different data types for similar statistical tests?
      - L28-31: If you are suggesting that in order to *access* the statistical details of objects from some packages, there are many data types, I think this is a reasonable statement — but you do need to specify this: e.g., "Statistical test objects returned by functions in packages often have information stored in multiple types (e.g., package X, package Y). By contrast, `statsExpressions` …. "
      - I suggest removing or completely redeveloping L32-36, it does not add to the manuscript substantially.

### Tidy Dataframes from Statistical Analysis

 In the Tidy section starting from L47, the examples demonstrate that effectively only a single string needs to be changed in the `statsExpressions::two_sample_test` function to use a parametric/robust t-test. This section can be simplified and reduced by removing the code examples (which verge on being API examples), and stating _directly in-line_ exactly what the functionality is. An example replacement could be: "A user can simply modify one argument in the `statsExpressions::two_sample_test` function to change between a robust, parametric or Bayesian analysis.".

The output code is not needed in the manuscript. If you believe it is crucial to show somewhere the dataframe format, I would dedicate a specific section and description to it, rather than show it for all functions by default.

You do not need full reproducible examples (setting seed; library loading) in the paper, particularly as these are already contained on the README.

I would also specify the `dplyr` grouped example in text (i.e., say it can be done, mention the function used), and point the reader to the vignettes or README where further details can be obtained.

### Expressions for Plots

 Figure 2 appears very useful and important, and immediately makes clear what the package offers. I suggest it appear first, with a description of the philosophy (i.e., gold standard), and then show the plot afterwards as an example output. You also can probably remove the generating code here, given the only unique argument is one string `df$expression[1]`.

### References

The APA reference is appearing incorrectly, you can [address this LaTeX](https://tex.stackexchange.com/questions/149769/how-to-cite-organizations-name-with-space-between-words) such that it treats American Psychological Association as a full singular name by wrapping in `{{name}}`. Also ensuring you specify which APA version (6 or 7),  rather than the original text from 1985.

## Other Feedback (related to package)

  - The consistent handling of missing data is a clear strength of the package. However, it is not made clear in outputs
    if or how much data was omitted due to this. I did wonder if it would be beneficial to have this information in the
    dataframe (or warning), as it seems particularly important for grouped analyses. That said, I also do not see this as a concern  precluding publication. I leave this to the digression of the author to decide.
  - I was curious as to why the test for support of missing data is omitted if R version \< 3.6? Just curious.
  - I also noticed that the package has incomplete support for `…` notation, meaning additional arguments to the underlying packages are not supported. This itself is not an issue, but I do think it would be worthwhile mentioning in the manuscript or documentation that some arguments (likely only to be used by for specific/advanced reasons) are not supported in the package (as a footnote perhaps).

-------------------

Please let me know if you have any questions or require clarification. Thank you for inviting me to review this work.
