### 2021-05-05 IndrajeetPatil/statsExpressions#94 @humanfactors
[JOSS Review] Documentation linking
## Functionality documentation

Related to JOSS submission: https://github.com/openjournals/joss-reviews/issues/3236

The functions in the package do callback, in some cases, through several other packages (e.g., `corr_test` > `correlation::correlation` > `BayesFactor::correlationBF`). This is not problematic in itself, but I have some concern regarding accessibility of the underlying arguments/functions.

Take for example, the prior specification in the correlation function. The `bf.prior` argument actually resolves to `rscale` in `BayesFactor::correlationBF`, which accepts string notation for the prior argument (e.g., "ultrawide"). Indeed, `statsExpression`  does support this prior string notation, but this is not documented or clear. Additionally, there is some added cognitive load of mapping between the different argument names (between `statsExpression` and underlying packages), which may arise if a user needs to find out more information about how an argument is resolved in the underlying package.

I suggest an approach to resolve this (ideally) would be for the in-built help documentation to link directly to the source of further information. It would be useful for each function to link to the underlying implementation. Roxygen comments can link to external packages: `\link[pkg]{function}`. This way, a user can quickly access further information about the implementation, without requiring external searching.

I appreciate you do provide a table that shows this online (and direct users there), so this particular issue is a suggestion rather than an urgent issue.
