### 2021-05-05 #3236 @humanfactors
I've now finished my review of StatsExpressions. I have added an issue to the original repository which includes my main feedback to the manuscript itself and items which should be resolved prior to acceptance (e.g., documentation updates). I have also opened a more specific issue, which I believe would dramatically improve the package, but I do not see as precluding publication.

- Manuscript Review: https://github.com/IndrajeetPatil/statsExpressions/issues/93
- Issue #1: https://github.com/IndrajeetPatil/statsExpressions/issues/94

I note that I believe this would be classed as a major utility package, which would fall within the scope of JOSS.

Thank you again, I enjoyed reviewing this work. You have developed a comprehensive package here.
