### 2021-05-05 IndrajeetPatil/statsExpressions#93 @humanfactors
Edited to be clear that `Only the last paragraph (L43-46)` was what compared packages.
