+++ paper/paper.Rmd
@@ -46,38 +46,31 @@ knitr::opts_chunk$set(
   comment = "#>"
 )
 
+# needed libraries
 library(dplyr)
 library(statsExpressions)
+library(ggplot2)
+library(ggridges)
+library(palmerpenguins)
+
+# for reproducibility
+set.seed(123) 
 ```
 
 # Summary
 
-The `statsExpressions` package is designed to facilitate producing dataframes
-with rich statistical details for the most common types of statistical
-approaches and tests: parametric, nonparametric, robust, and Bayesian *t*-test,
-one-way ANOVA, correlation analyses, contingency table analyses, and
-meta-analyses. The functions are pipe-friendly and provide a consistent syntax
-to work with tidy data. These dataframes additionally contain expressions with
-statistical details, and can be used in graphing packages to display these
-details.
+The `statsExpressions` package has two key aims: to provide a consistent syntax
+to do statistical analysis with tidy data, and to provide statistical
+expressions (i.e., a pre-formatted in-text statistical result) for plotting
+functions. Currently, it supports common types of statistical approaches and
+tests: parametric, nonparametric, robust, and Bayesian *t*-test, one-way ANOVA,
+correlation analyses, contingency table analyses, and meta-analyses. The
+functions are pipe-friendly and compatible with tidy data.
 
 # Statement of need
 
-The aim of this package is to provide an approachable and intuitive syntax to
-carry out common statistical tests across diverse statistical approaches.
-
-# Comparison to Other Packages
-
-Behind the scenes, `statsExpressions` uses `stats` package for parametric and
-non-parametric [@base2021], `WRS2` package for robust [@Mair2020], and
-`BayesFactor` package for Bayesian statistics [@Morey2020]. Additionally,
-random-effects meta-analysis is carried out using `metafor` (parametric)
-[@Viechtbauer2010], `metaplus` (robust) [@Beath2016], and `metaBMA` (Bayesian)
-[@Heck2019] packages. So one can naturally ask why there needs to be another
-package that wraps around these packages.
-
-There is a lot of diversity among these packages in terms of their syntax and
-expected input type that can make it difficult to switch from one statistical
+Statistical packages exhibit substantial diversity in terms of their syntax and
+expected input type. This can make it difficult to switch from one statistical
 approach to another. For example, some functions expect vectors as inputs, while
 others expect dataframes. Depending on whether it is a repeated measures design
 or not, different functions might expect data to be in wide or long format. Some
@@ -85,24 +78,27 @@ functions can internally omit missing values, while other functions error in
 their presence. Furthermore, if someone wishes to utilize the objects returned
 by these packages downstream in their workflow, this is not straightforward
 either because even functions from the same package can return a list, a matrix,
-an array, a dataframe, etc., depending on the function. So on and so forth.
-
-The result of sustained exposure to such inconsistencies is that data
-exploration can become a cognitively demanding task and discourage users to
-explore different statistical approaches. In the long run, this might even
-solidify into a habit of sticking to the defaults without giving much thought to
-the alternative approaches (e.g., exploring if Bayesian hypothesis testing is to
-be preferred over null hypothesis significance testing in context of the
-problem).
+an array, a dataframe, etc., depending on the function.
 
 This is where `statsExpressions` comes in: It can be thought of as a unified
 portal through which most of the functionality in these underlying packages can
-be accessed, with a little to no cognitive overhead. The package offers just six
[…]
