+++ paper/paper.bib
@@ -139,10 +139,12 @@
 }
 
 
-@book{american1985publication,
-  title={Publication Manual of the American Psychological Association: American Psychological Association},
+@book{american2019publication,
+  title={{Publication Manual of the American Psychological Association}},
+  edition = {7th Edition},
   author={American Psychological Association and others},
-  year={1985},
+  year={2019},
+  isbn={1433832178},
   publisher={American Psychological Association}
 }
 
