+++ DESCRIPTION
@@ -12,7 +12,7 @@ Maintainer: Indrajeet Patil <patilindrajeet.science@gmail.com>
 Description: Utilities for producing dataframes with rich details for the
     most common types of statistical approaches and tests: parametric,
     nonparametric, robust, and Bayesian t-test, one-way ANOVA, correlation
-    analyses, contingency table analyses, and meta-analyses.  The
+    analyses, contingency table analyses, and meta-analyses. The
     functions are pipe-friendly and provide a consistent syntax to work
     with tidy data. These dataframes additionally contain expressions with
     statistical details, and can be used in graphing packages. This

[…]
