+++ paper/paper.Rmd
@@ -15,7 +15,6 @@ affiliations:
   - name: Center for Humans and Machines, Max Planck Institute for Human Development, Berlin, Germany
     index: 1
 date: "`r Sys.Date()`"
-year: 2021
 bibliography: paper.bib
 output: rticles::joss_article
 csl: apa.csl
@@ -54,7 +53,7 @@ library(ggridges)
 library(palmerpenguins)
 
 # for reproducibility
-set.seed(123) 
+set.seed(123)
 ```
 
 # Summary
@@ -125,11 +124,10 @@ Additionally, it relies on `easystats` packages [@Ben-Shachar2020;
 appropriate effect size/posterior estimates and their confidence/credible
 intervals.
 
-To illustrate the simplicity of this syntax, let's say we want to compare
-equality of a measure among two independent groups. We can use the
-`two_sample_test` function here. If we first run a parametric *t*-test and then
-decide to run a robust *t*-test instead, the syntax remains the same and the
-statistical approach can be modified by changing a single argument:
+To illustrate the simplicity of this syntax, let's say we want to run a
+two-sample *t*-test. If we first run a parametric *t*-test and then decide to
+run a robust *t*-test instead, the syntax remains the same and the statistical
+approach can be modified by changing a single argument:
 
 ```{r df_p}
 mtcars %>% two_sample_test(am, wt, type = "parametric") # Welch's t-test
@@ -138,16 +136,7 @@ mtcars %>% two_sample_test(am, wt, type = "robust") # Yuen's t-test
 ```
 
 These functions are also compatible with other popular data manipulation
-packages. For example, we can use `dplyr` to repeat the same analysis across
-grouping variables.
-
-```{r grouped_df, echo=FALSE, eval=FALSE}
-# running one-sample proportion test for all levels of `cyl`
-mtcars %>%
-  group_by(cyl) %>%
-  group_modify(~ contingency_table(.x, am), .keep = TRUE) %>%
-  ungroup()
-```
+packages (see Appendix for an example).
 
 # Expressions for Plots
 
@@ -166,17 +155,32 @@ results in the context of a visualization is indeed a philosophy adopted by the
 `ggstatsplot` package [@Patil2021], and `statsExpressions` functions as its
 statistical processing backend.
 
-```{r anova_example, fig.cap="Example illustrating how `statsExpressions` functions can be used to display results from a statistical test in a plot."}
-# creating a dataframe (for the `penguins` dataset from `palmerpenguins` package)
+```{r anova_example, echo=FALSE, fig.cap="Example illustrating how `statsExpressions` functions can be used to display results from a statistical test in a plot. Code to create this figure is reported in Appendix."}
+# needed libraries
+library(statsExpressions)
+library(ggplot2)
+library(ggridges)
+library(palmerpenguins) # `penguins` dataset is from this package
+
+# creating a dataframe
 res <- oneway_anova(penguins, species, body_mass_g, type = "nonparametric")
 
 # create a ridgeplot using `ggridges` package
 ggplot(penguins, aes(x = body_mass_g, y = species)) +
   geom_density_ridges(
-    jittered_points = TRUE, quantile_lines = TRUE, scale = 0.9, vline_size = 1, 
-    vline_color = "red", position = position_raincloud(adjust_vlines = TRUE)
+    jittered_points = TRUE,
+    quantile_lines = TRUE,
+    scale = 0.9,
+    vline_size = 1,
+    vline_color = "red",
+    position = position_raincloud(adjust_vlines = TRUE)
   ) + # use 'expression' column to display results in the subtitle
[…]
