### 2021-05-05 IndrajeetPatil/statsExpressions#93 @IndrajeetPatil
_**Making one minor change: Adding `dplyr` example and code for Figure 2 in "Appendix" instead of completely removig it.**_
