### 2021-05-05 IndrajeetPatil/statsExpressions#94 @IndrajeetPatil
Thanks, @humanfactors.

I agree that I am currently not doing a great job of making it for users to read the documentation for underlying functions. Let me see what I can do.
