+++ R/meta_analysis.R
@@ -17,11 +17,13 @@
 #'
 #' A dataframe containing results from random-effects meta-analysis.
 #'
-#' For more details, see-
+#' To see details about functions which are internally used to carry out these
+#' analyses, see the following vignette-
 #' \url{https://indrajeetpatil.github.io/statsExpressions/articles/stats_details.html}
 #'
 #' @note **Important**: The function assumes that you have already downloaded
 #'   the needed package (`metafor`, `metaplus`, or `metaBMA`) for meta-analysis.
+#'   If they are not available, you will be asked to install them.
 #'
 #' @importFrom rlang exec !!! call2
 #'
