+++ R/contingency_table.R
@@ -6,7 +6,8 @@
 #' A dataframe containing results from for contingency table analysis or
 #' goodness of fit test.
 #'
-#' For more details, see-
+#' To see details about functions which are internally used to carry out these
+#' analyses, see the following vignette-
 #' \url{https://indrajeetpatil.github.io/statsExpressions/articles/stats_details.html}
 #'
 #' @param x The variable to use as the **rows** in the contingency table.
