+++ man/oneway_anova.Rd
@@ -17,7 +17,7 @@ oneway_anova(
   var.equal = FALSE,
   bf.prior = 0.707,
   tr = 0.2,
-  nboot = 100,
+  nboot = 100L,
   top.text = NULL,
   ...
 )
@@ -73,7 +73,11 @@ omega-squared).}
     (or Satterthwaite) approximation to the degrees of freedom is used.}
 
 \item{bf.prior}{A number between \code{0.5} and \code{2} (default \code{0.707}), the prior
-width to use in calculating Bayes factors and posterior estimates.}
+width to use in calculating Bayes factors and posterior estimates. In
+addition to numeric arguments, several named values are also recognized:
+\code{"medium"}, \code{"wide"}, and \code{"ultrawide"}, corresponding to \emph{r} scale values
+of 1/2, sqrt(2)/2, and 1, respectively. In case of an ANOVA, this value
+corresponds to scale for fixed effects.}
 
 \item{tr}{Trim level for the mean when carrying out \code{robust} tests. In case
 of an error, try reducing the value of \code{tr}, which is by default set to
@@ -90,7 +94,8 @@ mostly relevant in the context of \code{ggstatsplot} functions.}
 \description{
 A dataframe containing results for one-way ANOVA.
 
-For more details, see-
+To see details about functions which are internally used to carry out these
+analyses, see the following vignette-
 \url{https://indrajeetpatil.github.io/statsExpressions/articles/stats_details.html}
 }
 \note{
