### 2021-05-05 IndrajeetPatil/statsExpressions#94 @IndrajeetPatil
I have made one commit to address this by having the functions front and center on the README and in vignettes. I think that's more than enough to guide participants to the right function.

Hope that this suffices.

Thanks once again for the comments.
