### 2021-05-05 #3236 @IndrajeetPatil
@humanfactors Thanks a lot for your thoughtful feedback!

I have significantly modified the paper and slightly adjusted the package documentation to address your concerns.

I have closed both of the issues you mention. Do not hesitate to open additional issues if you feel there are other points that need to be addressed.
