### 2021-05-05 c42d8ba
just remove the gallery vignette
