+++ pkgdown.yml
@@ -4,8 +4,7 @@ pkgdown_sha: ~
 articles:
   stats_details: stats_details.html
   web_only/dataframe_outputs: dataframe_outputs.html
-  web_only/gallery: gallery.html
-last_built: 2021-05-05T18:25Z
+last_built: 2021-05-05T19:06Z
 urls:
   reference: https://indrajeetpatil.github.io/statsExpressions//reference
   article: https://indrajeetpatil.github.io/statsExpressions//articles
