### 2021-05-06 IndrajeetPatil/statsExpressions#93 @humanfactors
Hi @IndrajeetPatil, great job on addressing the comments above. I still have a minor issue remaining with the manuscript.

In referencing to providing a concrete example on L13-21, I think your reasoning is sensible and examples can be omitted (see your response below).

> **[L29-31]** Furthermore, if someone wishes to utilize the objects returned by these packages downstream in their workflow, this is not straightforward either because even functions from the same package can return a list, a matrix, an array, a dataframe, etc., depending on the function.

> _**[@IndrajeetPatil] For both of these points, I am hesitant to do this because I do not want to single out a particular function or a package or. a data type. The argument is that there is diversity among all of the things you have mentioned and this statement is trivially true. So I don't think it is necessary to provide specific examples of such differences since the users must have already observed this in their prior experience. But, if you insist, I can reconsider this.**_

However, I still believe the manuscript would benefit from addressing my second point ("_If you are suggesting that in order to..._"). To clarify this point, I was not necessarily stating that you need explicit examples; but that the text on L29-31 is unclear. My assumed interpretation is that: **In many packages, the statistical test objects returned by the test functions
do not have all required information (e.g., degrees of freedom, significance, Bayes factor) accessible in a consistent data type. Depending on the specific test object and statistic in question, details may returned as a list, a matrix, an array, or a dataframe.**

If my explanation above is correct, this is a different meaning than what is implied by saying that "functions from the same package can return different types", with the type "depending on the function", which implies either multiple function return values or multiple functions for a test object.

I believe it is important to revise and make this small sentence clear, because it is major strength of your package!

-------
**Further information**

Basically I am assuming what you are getting at, is that for a single object (let's take `BFoneSample` from `BayesFactor`), to even begin to extract the required details, your workflow would being looking like below, where there many data types required to get all object information:

```r
data(sleep)
myobject = BayesFactor::ttestBF(x = sleep$extra[sleep$group==1], y = sleep$extra[sleep$group==2], paired=TRUE)
bayes_factor = BayesFactor::extractBF(myobject)
sample_size = nrow(attributes(myobject)$data)
attributes(myobject)$bayesFactor[["bf"]]
ps_median = parameters::model_parameters(myobject)$Median
```

And once you factor in switching between multiple approaches (robust/parametric), it becomes completely unmanageable.
