### 2021-05-06 #3236 @humanfactors
@IndrajeetPatil Great revisions!

- I have made a pull request adjusting to double braces for APA reference, which should fix the issue.
- I have one minor concern remaining: https://github.com/IndrajeetPatil/statsExpressions/issues/93#issuecomment-833219105
- I'm still uncertain whether the appendix is required with the examples of function use, given that the material is already contained in the README/Vignettes. I believe that textual descriptions would be equally as compelling in demonstrating the simplicity. However, I'm not stuck on this issue, and would leave it to the @mikldk to determine, perhaps after hearing from @tomfaulkenberry

Pending above, I am happy to move forward to accept :)
