### 2021-05-06 IndrajeetPatil/statsExpressions#93 @IndrajeetPatil
Thanks, Michael, for pushing me to be clearer here and for you suggested rephrasing. This is what I have changed the paragraph to. Do you find this more reasonable?

> **_Statistical packages exhibit substantial diversity in terms of their syntax and
expected input and output data type. For example, some functions expect vectors
as inputs, while others expect dataframes. Depending on whether it is a repeated
measures design or not, functions from the same package might expect data to be
in wide or tidy format. Some functions can internally omit missing values, while
other functions error in their presence. Furthermore, the statistical test
objects returned by the test functions might not have all required information
(e.g., degrees of freedom, significance, Bayes factor, etc.) accessible in a
consistent data type. Depending on the specific test object and statistic in
question, details may be returned as a list, a matrix, an array, or a dataframe.
This diversity can make it difficult to easily access all needed information for
hypothesis testing and estimation, and to switch from one statistical approach
to another._**
