+++ paper/paper.Rmd
@@ -69,19 +69,23 @@ functions are pipe-friendly and compatible with tidy data.
 # Statement of need
 
 Statistical packages exhibit substantial diversity in terms of their syntax and
-expected input type. This can make it difficult to switch from one statistical
-approach to another. For example, some functions expect vectors as inputs, while
-others expect dataframes. Depending on whether it is a repeated measures design
-or not, different functions might expect data to be in wide or long format. Some
-functions can internally omit missing values, while other functions error in
-their presence. Furthermore, if someone wishes to utilize the objects returned
-by these packages downstream in their workflow, this is not straightforward
-either because even functions from the same package can return a list, a matrix,
-an array, a dataframe, etc., depending on the function.
+expected input and output data type. For example, some functions expect vectors
+as inputs, while others expect dataframes. Depending on whether it is a repeated
+measures design or not, functions from the same package might expect data to be
+in wide or tidy format. Some functions can internally omit missing values, while
+other functions error in their presence. Furthermore, the statistical test
+objects returned by the test functions might not have all required information
+(e.g., degrees of freedom, significance, Bayes factor, etc.) accessible in a
+consistent data type. Depending on the specific test object and statistic in
+question, details may be returned as a list, a matrix, an array, or a dataframe.
+This diversity can make it difficult to easily access all needed information for
+hypothesis testing and estimation, and to switch from one statistical approach
+to another.
 
 This is where `statsExpressions` comes in: It can be thought of as a unified
 portal through which most of the functionality in these underlying packages can
-be accessed, with a simpler interface and no requirement to change data format.
+be accessed, with a simpler interface and no requirement to change tidy data
+format.
 
 # Comparison to Other Packages
 
@@ -132,7 +136,7 @@ approach can be modified by changing a single argument:
 ```{r df_p}
 mtcars %>% two_sample_test(am, wt, type = "parametric") # Welch's t-test
 
-mtcars %>% two_sample_test(am, wt, type = "robust") # Yuen's t-test
+mtcars %>% two_sample_test(am, wt, type = "robust")     # Yuen's t-test
 ```
 
 These functions are also compatible with other popular data manipulation
