### 2021-05-06 #3236 @IndrajeetPatil
Thanks, @humanfactors, for following up on the revision.

I've accepted your PR to fix the citation issue and addressed the minor concern.

As for retaining `Appendix`, I think this is really important. I want the manuscript to be **_self-contained_** and asking the reader to go to README/vignette to see either the code or the `dplyr` example violates this requirement. Let's say, as an unfamiliar reader, I print this paper, get on the train without wi-fi, and see that it argues that it makes visualizing something really easy but just shows the visualization and asks me to go to the website to see the code. But I can't and this will be a major disappointment. I want to avoid situations like these. Does that make sense? 😅
