### 2021-05-06 #3236 @humanfactors
I actually agree with your logic here, but there is some ambiguity as to how much code can be presented without becoming "API documentation" (which should not be contained in the manuscript according to reviewer guidelines). In this instance, I take your point that it makes the manuscript self-contained and demonstrates the functionality, but I will defer to others (more experienced with JOSS) to make comment further 😊
