+++ DESCRIPTION
@@ -29,7 +29,7 @@ Imports:
     correlation (>= 0.6.1),
     dplyr,
     effectsize (>= 0.4.4-1),
-    insight (>= 0.13.0),
+    insight (>= 0.13.2),
     ipmisc,
     parameters (>= 0.13.0),
     performance (>= 0.7.1),
