### 2021-05-07 1b6f031
use insight function to check installation
