+++ NAMESPACE
@@ -56,6 +56,7 @@ importFrom(effectsize,kendalls_w)
 importFrom(effectsize,omega_squared)
 importFrom(effectsize,rank_biserial)
 importFrom(effectsize,rank_epsilon_squared)
+importFrom(insight,check_if_installed)
 importFrom(insight,format_value)
 importFrom(insight,standardize_names)
 importFrom(ipmisc,"%$%")
@@ -80,7 +81,6 @@ importFrom(rlang,call2)
 importFrom(rlang,enexpr)
 importFrom(rlang,enquo)
 importFrom(rlang,ensym)
-importFrom(rlang,eval_tidy)
 importFrom(rlang,exec)
 importFrom(rlang,expr)
 importFrom(rlang,new_formula)
