+++ NAMESPACE
@@ -25,13 +25,13 @@ importFrom(BayesFactor,anovaBF)
 importFrom(BayesFactor,contingencyTableBF)
 importFrom(BayesFactor,logMeanExpLogs)
 importFrom(BayesFactor,ttestBF)
+importFrom(WRS2,akp.effect)
 importFrom(WRS2,dep.effect)
 importFrom(WRS2,rmanova)
 importFrom(WRS2,t1way)
 importFrom(WRS2,trimse)
 importFrom(WRS2,wmcpAKP)
 importFrom(WRS2,yuen)
-importFrom(WRS2,yuen.effect.ci)
 importFrom(WRS2,yuend)
 importFrom(correlation,correlation)
 importFrom(dplyr,case_when)
