+++ DESCRIPTION
@@ -52,7 +52,6 @@ Suggests:
 VignetteBuilder: 
     knitr
 Config/testthat/edition: 3
-Config/testthat/parallel: true
 Encoding: UTF-8
 Language: en-US
 LazyData: true

[…]
