### 2021-05-09 52cd229
switch t-test example to ANOVA example
