### 2021-05-09 11ddd6f
use non pooled SD for d and g
