+++ R/helpers_expr_templates.R
@@ -170,27 +170,11 @@ expr_template <- function(data,
     # preparing expression
     expression <- substitute(
       expr = paste(
-        statistic.text,
-        " = ",
-        statistic,
-        ", ",
-        italic("p"),
-        " = ",
-        p.value,
-        ", ",
-        effsize.text,
-        " = ",
-        estimate,
-        ", CI"[conf.level],
-        " [",
-        estimate.LB,
-        ", ",
-        estimate.UB,
-        "]",
-        ", ",
-        n.text,
-        " = ",
-        n
+        statistic.text, " = ", statistic, ", ",
+        italic("p"), " = ", p.value, ", ",
+        effsize.text, " = ", estimate,
+        ", CI"[conf.level], " [", estimate.LB, ", ", estimate.UB, "], ",
+        n.text, " = ", n
       ),
       env = list(
         statistic.text = statistic.text,
@@ -215,29 +199,11 @@ expr_template <- function(data,
     # preparing expression
     expression <- substitute(
       expr = paste(
-        statistic.text,
-        "(",
-        parameter,
-        ") = ",
-        statistic,
-        ", ",
-        italic("p"),
-        " = ",
-        p.value,
-        ", ",
-        effsize.text,
-        " = ",
-        estimate,
-        ", CI"[conf.level],
-        " [",
-        estimate.LB,
-        ", ",
-        estimate.UB,
-        "]",
-        ", ",
-        n.text,
-        " = ",
-        n
+        statistic.text, "(", parameter, ") = ", statistic, ", ",
+        italic("p"), " = ", p.value, ", ",
+        effsize.text, " = ", estimate,
+        ", CI"[conf.level], " [", estimate.LB, ", ", estimate.UB, "], ",
+        n.text, " = ", n
       ),
       env = list(
         statistic.text = statistic.text,
@@ -261,31 +227,11 @@ expr_template <- function(data,
     # preparing expression
     expression <- substitute(
       expr = paste(
-        statistic.text,
-        "(",
-        parameter1,
-        ",",
-        parameter2,
-        ") = ",
-        statistic,
-        ", ",
-        italic("p"),
-        " = ",
-        p.value,
-        ", ",
-        effsize.text,
-        " = ",
-        estimate,
-        ", CI"[conf.level],
-        " [",
-        estimate.LB,
-        ", ",
-        estimate.UB,
-        "]",
-        ", ",
-        n.text,
-        " = ",
-        n
+        statistic.text, "(", parameter1, ",", parameter2, ") = ", statistic, ", ",
+        italic("p"), " = ", p.value, ", ",
+        effsize.text, " = ", estimate,
+        ", CI"[conf.level], " [", estimate.LB, ", ", estimate.UB, "], ",
+        n.text, " = ", n
       ),
       env = list(
         statistic.text = statistic.text,
