+++ tests/testthat/_snaps/oneway_anova_robust.md
@@ -22,7 +22,7 @@
       paste(italic("F")["trimmed-means"], "(", "2", ",", "18.97383", 
           ") = ", "20.24946", ", ", italic("p"), " = ", "2e-05", ", ", 
           widehat(xi), " = ", "0.85858", ", CI"["95%"], " [", "0.85268", 
-          ", ", "0.86448", "]", ", ", italic("n")["obs"], " = ", "32")
+          ", ", "0.86448", "], ", italic("n")["obs"], " = ", "32")
 
 ---
 
@@ -48,7 +48,7 @@
       paste(italic("F")["trimmed-means"], "(", "2", ",", "21.6869", 
           ") = ", "0.0503", ", ", italic("p"), " = ", "0.9511", ", ", 
           widehat(xi), " = ", "0.2013", ", CI"["95%"], " [", "0.0872", 
-          ", ", "0.7537", "]", ", ", italic("n")["obs"], " = ", "71")
+          ", ", "0.7537", "], ", italic("n")["obs"], " = ", "71")
 
 # expr_anova_robust works - within-subjects
 
@@ -77,6 +77,6 @@
       paste(italic("F")["trimmed-means"], "(", "2.7303", ",", "144.7051", 
           ") = ", "20.9752", ", ", italic("p"), " = ", "1.146e-10", 
           ", ", widehat(delta)["R-avg"]^"AKP", " = ", "0.6635", ", CI"["95%"], 
-          " [", "0.4660", ", ", "0.9707", "]", ", ", italic("n")["pairs"], 
+          " [", "0.4660", ", ", "0.9707", "], ", italic("n")["pairs"], 
           " = ", "88")
 
