+++ tests/testthat/_snaps/two_sample_robust.md
@@ -22,7 +22,7 @@
       paste(italic("t")["Yuen"], "(", "89", ") = ", "28.7230", ", ", 
           italic("p"), " = ", "0e+00", ", ", widehat(delta)["R"]^"AKP", 
           " = ", "2.3582", ", CI"["95%"], " [", "1.9615", ", ", "2.6081", 
-          "]", ", ", italic("n")["pairs"], " = ", "150")
+          "], ", italic("n")["pairs"], " = ", "150")
 
 # t_robust - within-subjects - with NAs
 
@@ -48,7 +48,7 @@
       paste(italic("t")["Yuen"], "(", "53", ") = ", "2.909", ", ", 
           italic("p"), " = ", "0.005", ", ", widehat(delta)["R"]^"AKP", 
           " = ", "0.410", ", CI"["95%"], " [", "0.238", ", ", "0.611", 
-          "]", ", ", italic("n")["pairs"], " = ", "90")
+          "], ", italic("n")["pairs"], " = ", "90")
 
 # t_robust - between-subjects - without NAs
 
@@ -74,7 +74,7 @@
       paste(italic("t")["Yuen"], "(", "13.584", ") = ", "5.840", ", ", 
           italic("p"), " = ", "4.85e-05", ", ", widehat(delta)["R"]^"AKP", 
           " = ", "2.482", ", CI"["95%"], " [", "0.738", ", ", "5.128", 
-          "]", ", ", italic("n")["obs"], " = ", "32")
+          "], ", italic("n")["obs"], " = ", "32")
 
 # t_robust - between-subjects - with NAs
 
@@ -100,5 +100,5 @@
       paste(italic("t")["Yuen"], "(", "13.8476", ") = ", "0.4521", 
           ", ", italic("p"), " = ", "0.6582", ", ", widehat(delta)["R"]^"AKP", 
           " = ", "-0.3583", ", CI"["95%"], " [", "-7.1637", ", ", "0.4061", 
-          "]", ", ", italic("n")["obs"], " = ", "29")
+          "], ", italic("n")["obs"], " = ", "29")
 
