### 2021-05-10 44c2852
change lintr and simplify tests
