+++ tests/testthat/test-meta_random_parametric.R
@@ -7,15 +7,14 @@ test_that(
     skip_if_not_installed("metafor")
 
     # dataframe
-    df_eg <-
-      structure(
-        list(
-          estimate = c(0.111, 0.245, 0.8, 1.1, 0.03),
-          std.error = c(0.05, 0.111, 0.001, 0.2, 0.01)
-        ),
-        row.names = c(NA, -5L),
-        class = c("tbl_df", "tbl", "data.frame")
-      )
+    df_eg <- structure(
+      list(
+        estimate = c(0.111, 0.245, 0.8, 1.1, 0.03),
+        std.error = c(0.05, 0.111, 0.001, 0.2, 0.01)
+      ),
+      row.names = c(NA, -5L),
+      class = c("tbl_df", "tbl", "data.frame")
+    )
 
     # dataframe output
     set.seed(123)
