### 2021-05-10 9b8fd9f
add additional args to one sample test function
