+++ R/one_sample_test.R
@@ -4,6 +4,22 @@
 #' @param x A numeric variable from the dataframe `data`.
 #' @param test.value A number indicating the true value of the mean (Default:
 #'   `0`).
+#' @param .f.args,.f.es.args A **list** of additional arguments to be passed to
+#'   internal functions `.f` and `.f.es`. These are used to carry out
+#'   statistical tests and to compute effect sizes, respectively. To see what
+#'   additional arguments are available, you can have a look at the
+#'   documentation for the respective function. Note that the chosen internal
+#'   functions (`.f` and `.f.es`) will themselves depend on the specified `type`
+#'   argument. The defaults reflect arguments relevant for parametric tests
+#'   because that's the default value for `type`. See `Details` section to see
+#'   which internal functions are used. The function might fail if you provide
+#'   an argument in a list which the underlying function does not take. This can
+#'   happen when you forget that you changed `type` argument, but forget to
+#'   change the `.f.args` and `.f.es.args` accordingly. Also, note that these
+#'   arguments are useful to provide *additional arguments*. Therefore, you
+#'   can's re-specify an argument you have already specified. For example, for
+#'   robust tests, you can use `tr` argument to specify trimming level, but then
+#'   you can't specify `tr` again inside list passed to `.f.args`
 #' @inheritParams ipmisc::long_to_wide_converter
 #' @inheritParams ipmisc::stats_type_switch
 #' @inheritParams expr_template
@@ -13,11 +29,26 @@
 #'
 #' @description
 #'
-#' A dataframe containing results from a one-sample test. The exact test and the
-#' effect size details contained will depend on the `type` argument.
+#' A dataframe containing results from a one-sample test.
 #'
-#' To see details about functions which are internally used to carry out these
-#' analyses, see the following vignette-
+#' @details
+#'
+#' The exact test and the effect size details contained will depend on the
+#' `type` argument.
+#'
+#'   Internal function `.f` used to carry out statistical test:
+#'   - **parametric**: `stats::t.test`
+#'   - **nonparametric**: `stats::wilcox.test`
+#'   - **robust**: `trimcibt` (custom)
+#'   - **bayes**: `BayesFactor::ttestBF`
+#'
+#'   Internal function `.f.es` used to compute effect size:
+#'   - **parametric**: `effectsize::cohens_d`, `effectsize::hedges_g`
+#'   - **nonparametric**: `effectsize::rank_biserial`
+#'   - **robust**: `trimcibt` (custom)
+#'   - **bayes**: `bayestestR::describe_posterior`
+#'
+#' For more, see-
 #' \url{https://indrajeetpatil.github.io/statsExpressions/articles/stats_details.html}
 #'
 #' @importFrom dplyr select mutate pull rename_all recode
@@ -82,7 +113,8 @@ one_sample_test <- function(data,
                             tr = 0.2,
                             bf.prior = 0.707,
                             effsize.type = "g",
-                            nboot = 100L,
+                            .f.args = list(),
+                            .f.es.args = list(verbose = FALSE, ci = conf.level),
                             top.text = NULL,
                             ...) {
   # standardize the type of statistics
@@ -112,18 +144,12 @@ one_sample_test <- function(data,
   # preparing expression
[…]
