+++ tests/testthat/test-one_two_sample_dataframes.R
@@ -2,6 +2,7 @@ test_that(
   desc = " parametric t-tests",
   code = {
     skip_if(getRversion() < "4.0")
+    options(tibble.width = Inf)
 
     # one-sample t-test (with NAs) ---------
 
