+++ R/one_sample_test.R
@@ -4,22 +4,8 @@
 #' @param x A numeric variable from the dataframe `data`.
 #' @param test.value A number indicating the true value of the mean (Default:
 #'   `0`).
-#' @param .f.args,.f.es.args A **list** of additional arguments to be passed to
-#'   internal functions `.f` and `.f.es`. These are used to carry out
-#'   statistical tests and to compute effect sizes, respectively. To see what
-#'   additional arguments are available, you can have a look at the
-#'   documentation for the respective function. Note that the chosen internal
-#'   functions (`.f` and `.f.es`) will themselves depend on the specified `type`
-#'   argument. The defaults reflect arguments relevant for parametric tests
-#'   because that's the default value for `type`. See `Details` section to see
-#'   which internal functions are used. The function might fail if you provide
-#'   an argument in a list which the underlying function does not take. This can
-#'   happen when you forget that you changed `type` argument, but forget to
-#'   change the `.f.args` and `.f.es.args` accordingly. Also, note that these
-#'   arguments are useful to provide *additional arguments*. Therefore, you
-#'   can's re-specify an argument you have already specified. For example, for
-#'   robust tests, you can use `tr` argument to specify trimming level, but then
-#'   you can't specify `tr` again inside list passed to `.f.args`
+#' @param effsize.type Type of effect size needed for *parametric* tests. The
+#'   argument can be `"d"` (for Cohen's *d*) or `"g"` (for Hedge's *g*).
 #' @inheritParams ipmisc::long_to_wide_converter
 #' @inheritParams ipmisc::stats_type_switch
 #' @inheritParams expr_template
@@ -113,8 +99,6 @@ one_sample_test <- function(data,
                             tr = 0.2,
                             bf.prior = 0.707,
                             effsize.type = "g",
-                            .f.args = list(),
-                            .f.es.args = list(verbose = FALSE, ci = conf.level),
                             top.text = NULL,
                             ...) {
   # standardize the type of statistics
@@ -144,12 +128,18 @@ one_sample_test <- function(data,
   # preparing expression
   if (type %in% c("parametric", "nonparametric")) {
     # extracting test details
-    stats_df <- rlang::exec(.f, x = x_vec, mu = test.value, !!!.f.args) %>%
+    stats_df <- rlang::exec(.f, x = x_vec, mu = test.value) %>%
       tidy_model_parameters(.) %>%
       dplyr::select(-dplyr::matches("^est|^conf|^diff|^term|^ci"))
 
     # extracting effect size details
-    effsize_df <- rlang::exec(.f.es, x = x_vec, mu = test.value, !!!.f.es.args) %>%
+    effsize_df <- rlang::exec(
+        .f.es,
+        x = x_vec,
+        mu = test.value,
+        verbose = FALSE,
+        ci = conf.level
+      ) %>%
       tidy_model_effectsize(.)
 
     # these can be really big values
@@ -164,7 +154,7 @@ one_sample_test <- function(data,
   if (type == "robust") {
     # bootstrap-t method for one-sample test
     no.parameters <- 0L
-    stats_df <- rlang::exec(trimcibt, x = x_vec, nv = test.value, tr = tr, !!!.f.es.args)
+    stats_df <- rlang::exec(trimcibt, x = x_vec, nv = test.value, tr = tr, ci = conf.level)
   }
 
   # expression
@@ -196,7 +186,7 @@ one_sample_test <- function(data,
 #' @importFrom WRS2 trimse
 #' @noRd
 
[…]
