+++ README.Rmd
@@ -152,9 +152,19 @@ Between group/condition comparisons | ✅ | ✅ | ✅ | ✅
 Within group/condition comparisons | ✅ | ✅ | ✅ | ✅
 Distribution of a numeric variable | ✅ | ✅ | ✅ | ✅
 Correlation between two variables | ✅ | ✅ | ✅ | ✅
-Association between categorical variables | ✅ | ✅ | `NA` | ✅
-Equal proportions for categorical variable levels | ✅ | ✅ | `NA` | ✅
-Random-effects meta-analysis | ✅ | `NA` | ✅ | ✅
+Association between categorical variables | ✅ | ✅ | ❌ | ✅
+Equal proportions for categorical variable levels | ✅ | ✅ | ❌ | ✅
+Random-effects meta-analysis | ✅ | ❌ | ✅ | ✅
+
+Summary of Bayesian analysis
+
+Analysis | Hypothesis testing | Estimation
+------------------ | ---------- | ---------
+(one/two-sample) *t*-test | ✅ | ✅
+one-way ANOVA | ✅ | ✅
+correlation | ✅ | ✅
+(one/two-way) contingency table | ✅ | ✅
+random-effects meta-analysis | ✅ | ✅
 
 # Statistical reporting
 
@@ -505,7 +515,7 @@ No. of groups: `> 2` => `oneway_anova`
 
 ### between-subjects
 
-Following (between-subjects) tests are carried out for each type of analyses-
+**Hypothesis testing**
 
 Type | No. of groups | Test | Function used
 ----------- | --- | ------------------------- | -----
@@ -518,15 +528,14 @@ Non-parametric | 2 | Mann–Whitney *U* test | `stats::wilcox.test`
 Robust | 2 |  Yuen's test for trimmed means | `WRS2::yuen`
 Bayesian | 2 | Student's *t*-test | `BayesFactor::ttestBF`
 
-Following effect sizes (and confidence intervals/CI) are available for each type
-of test-
+**Effect size estimation**
 
 Type | No. of groups | Effect size | CI? | Function used
 ----------- | --- | ------------------------- | --- | -----
 Parametric | > 2 | $\eta_{p}^2$, $\omega_{p}^2$ | ✅ | `effectsize::omega_squared`, `effectsize::eta_squared`
 Non-parametric | > 2 | $\epsilon_{ordinal}^2$ | ✅ | `effectsize::rank_epsilon_squared`
 Robust | > 2 | $\xi$ (Explanatory measure of effect size) | ✅ | `WRS2::t1way`
-Bayes Factor | > 2 | $R_{posterior}^2$ | ✅ | `performance::r2_bayes`
+Bayes Factor | > 2 | $R_{Bayesian}^2$ | ✅ | `performance::r2_bayes`
 Parametric | 2 | Cohen's *d*, Hedge's *g* | ✅ | `effectsize::cohens_d`, `effectsize::hedges_g`
 Non-parametric | 2 | *r* (rank-biserial correlation) | ✅ | `effectsize::rank_biserial`
 Robust | 2 |  $\delta_{R}^{AKP}$ (Algina-Keselman-Penfield robust standardized difference) | ✅ | `WRS2::akp.effect`
@@ -534,7 +543,7 @@ Bayesian | 2 | $\delta_{posterior}$ | ✅ | `bayestestR::describe_posterior`
 
 ### within-subjects
 
-Following (within-subjects) tests are carried out for each type of analyses-
+**Hypothesis testing**
 
 Type | No. of groups | Test | Function used
 ----------- | --- | ------------------------- | -----
[…]

[…]
