### 2021-05-11 5be8f1d
minor edits to docs
