+++ .github/workflows/check-full.yaml
@@ -24,7 +24,7 @@ jobs:
           - {os: windows-latest, r: 'devel', args: "--no-manual"}
           - {os: windows-latest, r: 'oldrel', args: "--no-manual"}
           - {os: macOS-latest, r: 'release', args: "--no-manual"}
-          #- {os: macOS-latest, r: 'devel', args: "--no-manual"}
+          - {os: macOS-latest, r: 'devel', args: "--no-manual"}
           - {os: macOS-latest, r: 'oldrel', args: "--no-manual"}
           - {os: ubuntu-16.04, r: 'release', cran: "https://demo.rstudiopm.com/all/__linux__/xenial/latest", args: "--no-manual"}
           - {os: ubuntu-16.04, r: 'devel', cran: "https://demo.rstudiopm.com/all/__linux__/xenial/latest", args: "--no-manual"}
