+++ tests/testthat/_snaps/two_sample_parametric.md
@@ -3,13 +3,16 @@
     Code
       dplyr::select(df1, -expression)
     Output
-      # A tibble: 1 x 13
+      # A tibble: 1 x 15
         term  group mean.group1 mean.group2 statistic df.error p.value
         <chr> <chr>       <dbl>       <dbl>     <dbl>    <dbl>   <dbl>
       1 len   supp         20.7        17.0      1.92       58  0.0604
         method            estimate conf.level conf.low conf.high effectsize
         <chr>                <dbl>      <dbl>    <dbl>     <dbl> <chr>     
       1 Two Sample t-test    0.495       0.99   -0.184      1.17 Cohen's d 
+        conf.method conf.distribution
+        <chr>       <chr>            
+      1 ncp         t                
 
 ---
 
@@ -26,13 +29,16 @@
     Code
       dplyr::select(df1, -expression)
     Output
-      # A tibble: 1 x 13
+      # A tibble: 1 x 15
         term  group mean.group1 mean.group2 statistic df.error p.value
         <chr> <chr>       <dbl>       <dbl>     <dbl>    <dbl>   <dbl>
       1 len   supp         20.7        17.0      1.92     55.3  0.0606
         method                  estimate conf.level conf.low conf.high effectsize
         <chr>                      <dbl>      <dbl>    <dbl>     <dbl> <chr>     
       1 Welch Two Sample t-test    0.488        0.9   0.0599     0.911 Hedges' g 
+        conf.method conf.distribution
+        <chr>       <chr>            
+      1 ncp         t                
 
 ---
 
@@ -49,13 +55,13 @@
     Code
       dplyr::select(df1, -expression)
     Output
-      # A tibble: 1 x 11
+      # A tibble: 1 x 13
         term  group     statistic df.error  p.value method        estimate conf.level
         <chr> <chr>         <dbl>    <dbl>    <dbl> <chr>            <dbl>      <dbl>
       1 value condition      34.8      149 1.85e-73 Paired t-test     2.83        0.5
-        conf.low conf.high effectsize
-           <dbl>     <dbl> <chr>     
-      1     2.71      2.96 Hedges' g 
+        conf.low conf.high effectsize conf.method conf.distribution
+           <dbl>     <dbl> <chr>      <chr>       <chr>            
+      1     2.71      2.96 Hedges' g  ncp         t                
 
 ---
 
@@ -72,13 +78,13 @@
     Code
       dplyr::select(df1, -expression)
     Output
-      # A tibble: 1 x 11
+      # A tibble: 1 x 13
         term   group     statistic df.error  p.value method        estimate conf.level
         <chr>  <chr>         <dbl>    <dbl>    <dbl> <chr>            <dbl>      <dbl>
       1 desire condition      3.61       89 0.000500 Paired t-test    0.381       0.95
-        conf.low conf.high effectsize
-           <dbl>     <dbl> <chr>     
-      1    0.167     0.597 Cohen's d 
+        conf.low conf.high effectsize conf.method conf.distribution
+           <dbl>     <dbl> <chr>      <chr>       <chr>            
+      1    0.167     0.597 Cohen's d  ncp         t                
 
 ---
 
