+++ tests/testthat/_snaps/two_sample_parametric.md
@@ -3,16 +3,16 @@
     Code
       dplyr::select(df1, -expression)
     Output
-      # A tibble: 1 x 15
+      # A tibble: 1 x 16
         term  group mean.group1 mean.group2 statistic df.error p.value
         <chr> <chr>       <dbl>       <dbl>     <dbl>    <dbl>   <dbl>
       1 len   supp         20.7        17.0      1.92       58  0.0604
-        method            estimate conf.level conf.low conf.high effectsize
-        <chr>                <dbl>      <dbl>    <dbl>     <dbl> <chr>     
-      1 Two Sample t-test    0.495       0.99   -0.184      1.17 Cohen's d 
-        conf.method conf.distribution
-        <chr>       <chr>            
-      1 ncp         t                
+        method            alternative estimate conf.level conf.low conf.high
+        <chr>             <chr>          <dbl>      <dbl>    <dbl>     <dbl>
+      1 Two Sample t-test two.sided      0.495       0.99   -0.184      1.17
+        effectsize conf.method conf.distribution
+        <chr>      <chr>       <chr>            
+      1 Cohen's d  ncp         t                
 
 ---
 
@@ -29,16 +29,16 @@
     Code
       dplyr::select(df1, -expression)
     Output
-      # A tibble: 1 x 15
+      # A tibble: 1 x 16
         term  group mean.group1 mean.group2 statistic df.error p.value
         <chr> <chr>       <dbl>       <dbl>     <dbl>    <dbl>   <dbl>
       1 len   supp         20.7        17.0      1.92     55.3  0.0606
-        method                  estimate conf.level conf.low conf.high effectsize
-        <chr>                      <dbl>      <dbl>    <dbl>     <dbl> <chr>     
-      1 Welch Two Sample t-test    0.488        0.9   0.0599     0.911 Hedges' g 
-        conf.method conf.distribution
-        <chr>       <chr>            
-      1 ncp         t                
+        method                  alternative estimate conf.level conf.low conf.high
+        <chr>                   <chr>          <dbl>      <dbl>    <dbl>     <dbl>
+      1 Welch Two Sample t-test two.sided      0.488        0.9   0.0599     0.911
+        effectsize conf.method conf.distribution
+        <chr>      <chr>       <chr>            
+      1 Hedges' g  ncp         t                
 
 ---
 
@@ -55,13 +55,13 @@
     Code
       dplyr::select(df1, -expression)
     Output
-      # A tibble: 1 x 13
-        term  group     statistic df.error  p.value method        estimate conf.level
-        <chr> <chr>         <dbl>    <dbl>    <dbl> <chr>            <dbl>      <dbl>
-      1 value condition      34.8      149 1.85e-73 Paired t-test     2.83        0.5
-        conf.low conf.high effectsize conf.method conf.distribution
-           <dbl>     <dbl> <chr>      <chr>       <chr>            
-      1     2.71      2.96 Hedges' g  ncp         t                
+      # A tibble: 1 x 14
+        term  group     statistic df.error  p.value method        alternative estimate
+        <chr> <chr>         <dbl>    <dbl>    <dbl> <chr>         <chr>          <dbl>
+      1 value condition      34.8      149 1.85e-73 Paired t-test two.sided       2.83
+        conf.level conf.low conf.high effectsize conf.method conf.distribution
+             <dbl>    <dbl>     <dbl> <chr>      <chr>       <chr>            
+      1        0.5     2.71      2.96 Hedges' g  ncp         t                
 
 ---
 
@@ -78,13 +78,16 @@
     Code
       dplyr::select(df1, -expression)
     Output
-      # A tibble: 1 x 13
[…]
