### 2021-05-19 #3236 @IndrajeetPatil
@mikldk In the reduced working mode of JOSS, the reviews are supposed to be finished under 6 weeks?
Just anticipating how much longer this process will take.
