### 2021-05-19 #3236 @tomfaulkenberry
@IndrajeetPatil I just completed my review of the software.  Everything installed just fine. The sample analyses given in the software paper (and on the Github repo) work exactly as described on my machine. The syntax is intuitive and easy to customize.

One quick thing I wanted to ask about -- I wonder why Bayes factors are reported "raw" in the outputs, but with log(BF) in the plot expressions. Obviously, taking the log is useful in many cases (especially when the Bayes factors are really big/small.  Just curious about this choice.

Thanks to the excellent review by @humanfactors (and the corresponding updates thus performed), I found everything to be ready to publish as is.  Thanks for inviting me to review and learn about this neat package!

Tom
