### 2021-05-19 #3236 @IndrajeetPatil
@tomfaulkenberry Thanks a ton for the quick review and your kind words about the software.

> One quick thing I wanted to ask about -- I wonder why Bayes factors are reported "raw" in the outputs, but with log(BF) in the plot expressions. Obviously, taking the log is useful in many cases (especially when the Bayes factors are really big/small. Just curious about this choice.

In addition to this benefit you mentioned, the expressions have the additional restriction that they need to fit in the plot dimensions. The raw Bayes Factor like, e.g., `34547` will occupy more space than logged BF, which will be `10`. This is important for packages like `ggstatsplot` (https://indrajeetpatil.github.io/ggstatsplot/) that use these expressions to display results in a plot because then the length of the expression remains manageable.

![image](https://user-images.githubusercontent.com/11330453/118806493-7102f300-b8a7-11eb-8d16-b6d14309b75d.png)
