### 2021-05-20 #3236 @mikldk
@IndrajeetPatil:

* Please make the title at Zenodo the same as the paper's.
* There is some confusing about the version: Zenodo has a 1.1.0 prior to this 1.0.1 -- this is [not standard](https://en.wikipedia.org/wiki/Software_versioning); can you please fix it.
