### 2021-05-20 #3236 @IndrajeetPatil
@mikldk Thanks for pointing these inconsistencies:

> Please make the title at Zenodo the same as the paper's.

Done! The new title is the same as the paper
https://zenodo.org/record/4773886#.YKZOk5MzZQI

> There is some confusing about the version: Zenodo has a 1.1.0 prior to this 1.0.1 -- this is not standard; can you please fix it.

Good catch! This was a mistake on my part.
The previous release on zenodo was supposed to be `0.1.0`, but I apparently typed in `1.1.0` instead. I have fixed this mistake. So now the versions should make more sense.
Note that this mistake is only present on zenodo, but neither on CRAN nor on GitHub versions.

Let me know if there is anything else to fix.
