### 2021-05-20 #3236 @mikldk
@IndrajeetPatil:

Zenodo:
statsExpressions: Tidy Dataframes and Expressions with Statistical Details

JOSS:
statsExpressions: R Package for Tidy Dataframes and1Expressions with Statistical Details
